# Conflicts:
#	d360.web/Scripts/app/components/assignments/assignment-list/assignment-list.component.ts

+++ d360.core/entities/Metric/MetricApiViewModels.cs
@@ -691,13 +691,13 @@ namespace d360.core.entities.Metric
 
         public DateTime? RunDate { get; set; }
 
-        public int? TotalCount { get; set; }
+        public long? TotalCount { get; set; }
 
         public float PassFraction { get; set; }
 
-        public int? PassCount { get; set; }
+        public long? PassCount { get; set; }
 
-        public int? FailCount { get; set; }
+        public long? FailCount { get; set; }
     }
 
     public class DataQualityScoreItemEvidenceItemRollupPathViewModel
